% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salReturn.R
\name{salReturnUI}
\alias{salReturnUI}
\title{销售订单数据同步任务生成生成器界面}
\usage{
salReturnUI(
  tabTitle = "BYD销售退货同步",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = salReturnUI_left,
  func_right = salReturnUI_right,
  func_bottom = salReturnUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
销售订单数据同步任务生成生成器界面
}
\examples{
salReturnUI()
}
